acc_social/requirements.txt
django
djangorestframework
django-cors-headers
django-admin-inline-paginator